TP2/requirements.txt
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
torch>=2.0.0
pillow>=10.0.0
streamlit>=1.32.0
